# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "seo-content-agent",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/keywords": {
      "post": {
        "tags": [
          "keywords"
        ],
        "summary": "Create Keyword",
        "operationId": "create_keyword_api_v1_projects__project_id__keywords_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeywordCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KeywordDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "keywords"
        ],
        "summary": "List Keywords",
        "operationId": "list_keywords_api_v1_projects__project_id__keywords_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "intent",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Intent"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "min_volume",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Volume"
            }
          },
          {
            "name": "max_volume",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Volume"
            }
          },
          {
            "name": "min_difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Difficulty"
            }
          },
          {
            "name": "max_difficulty",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Difficulty"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KeywordList_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/keywords/{keyword_id}": {
      "get": {
        "tags": [
          "keywords"
        ],
        "summary": "Get Keyword",
        "operationId": "get_keyword_api_v1_projects__project_id__keywords__keyword_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "keyword_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Keyword Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KeywordDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "keywords"
        ],
        "summary": "Update Keyword",
        "operationId": "update_keyword_api_v1_projects__project_id__keywords__keyword_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "keyword_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Keyword Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeywordUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KeywordDetail_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/keywords/import": {
      "post": {
        "tags": [
          "keywords"
        ],
        "summary": "Import Keywords",
        "operationId": "import_keywords_api_v1_projects__project_id__keywords_import_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_keywords_api_v1_projects__project_id__keywords_import_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KeywordImportResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/keywords/imports/{import_id}": {
      "get": {
        "tags": [
          "keywords"
        ],
        "summary": "Get Import Status",
        "operationId": "get_import_status_api_v1_projects__project_id__keywords_imports__import_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          },
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_KeywordImportResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse_KeywordDetail_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KeywordDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[KeywordDetail]"
      },
      "ApiResponse_KeywordImportResponse_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KeywordImportResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[KeywordImportResponse]"
      },
      "ApiResponse_KeywordList_": {
        "properties": {
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KeywordList"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "type": "array",
            "title": "Errors"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[KeywordList]"
      },
...
      "Body_import_keywords_api_v1_projects__project_id__keywords_import_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_keywords_api_v1_projects__project_id__keywords_import_post"
      },
...
      "ErrorItem": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "details": {
            "type": "object",
            "title": "Details"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorItem",
        "description": "Stable machine-readable error entry."
      },
      "FunnelStage": {
        "type": "string",
        "enum": [
          "TOFU",
          "MOFU",
          "BOFU"
        ],
        "title": "FunnelStage"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeywordCreate": {
        "properties": {
          "keyword": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Keyword"
          },
          "search_volume": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Search Volume",
            "default": 0
          },
          "keyword_difficulty": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Keyword Difficulty",
            "default": 0.0
          },
          "cpc": {
            "type": "number",
            "minimum": 0.0,
            "title": "Cpc",
            "default": 0.0
          },
          "intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchIntent"
              },
              {
                "type": "string"
              }
            ],
            "title": "Intent",
            "default": "UNKNOWN"
          },
          "funnel_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FunnelStage"
              },
              {
                "type": "string"
              }
            ],
            "title": "Funnel Stage",
            "default": "TOFU"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KeywordSource"
              },
              {
                "type": "string"
              }
            ],
            "title": "Source",
            "default": "MANUAL"
          },
          "website_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "keyword"
        ],
        "title": "KeywordCreate"
      },
      "KeywordDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "website_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Website Id"
          },
          "keyword": {
            "type": "string",
            "title": "Keyword"
          },
          "normalized_keyword": {
            "type": "string",
            "title": "Normalized Keyword"
          },
          "search_volume": {
            "type": "integer",
            "title": "Search Volume"
          },
          "keyword_difficulty": {
            "type": "number",
            "title": "Keyword Difficulty"
          },
          "cpc": {
            "type": "number",
            "title": "Cpc"
          },
          "intent": {
            "type": "string",
            "title": "Intent"
          },
          "intent_confidence": {
            "type": "number",
            "title": "Intent Confidence"
          },
          "funnel_stage": {
            "type": "string",
            "title": "Funnel Stage"
          },
          "business_value_score": {
            "type": "number",
            "title": "Business Value Score"
          },
          "priority_score": {
            "type": "number",
            "title": "Priority Score"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "provider_metadata": {
            "type": "object",
            "title": "Provider Metadata"
          },
          "cluster_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Id"
          },
          "cluster_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cluster Name"
          },
          "target_page_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Page Id"
          },
          "target_page_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Page Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "project_id",
          "keyword",
          "normalized_keyword",
          "search_volume",
          "keyword_difficulty",
          "cpc",
          "intent",
          "intent_confidence",
          "funnel_stage",
          "business_value_score",
          "priority_score",
          "source",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "KeywordDetail"
      },
      "KeywordImportResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "title": "Organization Id"
          },
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "valid_rows": {
            "type": "integer",
            "title": "Valid Rows"
          },
          "invalid_rows": {
            "type": "integer",
            "title": "Invalid Rows"
          },
          "duplicate_rows": {
            "type": "integer",
            "title": "Duplicate Rows"
          },
          "created_rows": {
            "type": "integer",
            "title": "Created Rows"
          },
          "updated_rows": {
            "type": "integer",
            "title": "Updated Rows"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "project_id",
          "filename",
          "source",
          "total_rows",
          "valid_rows",
          "invalid_rows",
          "duplicate_rows",
          "created_rows",
          "updated_rows",
          "status",
          "created_at"
        ],
        "title": "KeywordImportResponse"
      },
      "KeywordList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/KeywordDetail"
            },
            "type": "array",
            "title": "Items"
          },
          "total_count": {
            "type": "integer",
            "title": "Total Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "KeywordList"
      },
...
      "KeywordSource": {
        "type": "string",
        "enum": [
          "MANUAL",
          "CSV",
          "SEARCH_CONSOLE",
          "KEYWORD_PROVIDER",
          "AI_DISCOVERY",
          "IMPORT"
        ],
        "title": "KeywordSource"
      },
      "KeywordStatus": {
        "type": "string",
        "enum": [
          "active",
          "archived",
          "ignored"
        ],
        "title": "KeywordStatus"
      },
      "KeywordUpdate": {
        "properties": {
          "search_volume": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Volume"
          },
          "keyword_difficulty": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Keyword Difficulty"
          },
          "cpc": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cpc"
          },
          "intent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SearchIntent"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "funnel_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FunnelStage"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Funnel Stage"
          },
          "business_value_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Value Score"
          },
          "priority_score": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Priority Score"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KeywordStatus"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "KeywordUpdate"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has More"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id"
        ],
        "title": "ResponseMeta",
        "description": "Request metadata shared by success and failure responses."
      },
...
      "SearchIntent": {
        "type": "string",
        "enum": [
          "INFORMATIONAL",
          "COMMERCIAL",
          "TRANSACTIONAL",
          "NAVIGATIONAL",
          "LOCAL",
          "UNKNOWN"
        ],
        "title": "SearchIntent"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}